refactor: 알림 스와이프 클래스 분리
- 뷰에서 있던 코드의 스와이프 기능 역할 분리
